+++ Compiler/Features/LanguageServer/UseCase/Completion/CompletionInteractor.cs
@@ -5,10 +5,9 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
-using KSPCompiler.Features.LanguageServer.UseCase.Abstractions;
 using KSPCompiler.Features.LanguageServer.UseCase.Abstractions.Completion;
+using KSPCompiler.Features.LanguageServer.UseCase.Completion.CompletionItems;
 using KSPCompiler.Shared.Domain.Compilation.Symbols;
-using KSPCompiler.Shared.Domain.Compilation.Symbols.MetaData.Extensions;
 
 namespace KSPCompiler.Features.LanguageServer.UseCase.Completion;
 
@@ -21,11 +20,14 @@ public sealed class CompletionInteractor : ICompletionUseCase
             var compilerCacheService = parameter.Input.Cache;
             var scriptLocation = parameter.Input.Location;
             var position = parameter.Input.Position;
+            var preferSnippetInsertion = parameter.Input.PreferSnippetInsertion;
 
             var cache = compilerCacheService.GetCache( scriptLocation );
             var symbolTable = cache.SymbolTable;
             var word = DocumentUtility.ExtractWord( cache.AllLinesText, position );
             var completions = new List<CompletionItem>();
+
+            var snippetTextBuilder = new StringBuilder();
 
             #region Collection of target symbols
             // プリプロセッサ
@@ -77,14 +79,22 @@ public sealed class CompletionInteractor : ICompletionUseCase
             #endregion ~Collection of target symbols
 
             #region Build completion list
-            BuildCompletionItem( preprocessors, word, CompletionItemKind.Variable, "Preprocessor", completions );
-            BuildCompletionItem( pgsKeyIds, word, CompletionItemKind.Variable, "PGS Key-Id", completions );
-            BuildCompletionItem( userVariables, word, CompletionItemKind.Variable, "User Variable", completions );
-            BuildCompletionItem( builtInVariables, word, CompletionItemKind.Function, "Built-in Variable", completions );
-            BuildCompletionItem( uiTypes, word, CompletionItemKind.Class, "UI Type", completions );
-            BuildCompletionItem( commands, word, CompletionItemKind.Method, "Command", completions );
-            BuildCompletionItem( userFunctions, word, CompletionItemKind.Function, "User Function", completions );
-            BuildCompletionItem( builtInCallBacks, word, CompletionItemKind.Event, "Callback", completions );
+            var preprocessorItemFactory = new PreprocessorCompletionItemFactory();
+            var pgsItemFactory = new PgsKeyCompletionItemFactory();
+            var variableItemFactory = new VariableCompletionItemFactory();
+            var uiItemFactory = new UICompletionItemFactory( snippetTextBuilder );
+            var commandItemFactory = new CommandCompletionItemFactory( snippetTextBuilder );
+            var callbackItemFactory = new CallbackCompletionItemFactory( snippetTextBuilder );
+            var userFunctionItemFactory = new UserFunctionCompletionItemFactory( snippetTextBuilder );
+
+            BuildCompletionItemNew( preprocessors, preprocessorItemFactory, word, preferSnippetInsertion, completions );
+            BuildCompletionItemNew( pgsKeyIds, pgsItemFactory, word, preferSnippetInsertion, completions );
+            BuildCompletionItemNew( userVariables, variableItemFactory, word, preferSnippetInsertion, completions );
+            BuildCompletionItemNew( builtInVariables, variableItemFactory, word, preferSnippetInsertion, completions );
+            BuildCompletionItemNew( uiTypes, uiItemFactory, word, preferSnippetInsertion, completions );
+            BuildCompletionItemNew( commands, commandItemFactory, word, preferSnippetInsertion, completions );
+            BuildCompletionItemNew( userFunctions, userFunctionItemFactory, word, preferSnippetInsertion, completions );
+            BuildCompletionItemNew( builtInCallBacks, callbackItemFactory, word, preferSnippetInsertion, completions );
             #endregion ~Build completion list
 
             await Task.CompletedTask;
@@ -126,202 +136,23 @@ public sealed class CompletionInteractor : ICompletionUseCase
         return list.Select( x => x.First().Value ).ToList();
     }
 
-    private static void BuildCompletionItem<TSymbol>(
+    private static void BuildCompletionItemNew<TSymbol>(
         IReadOnlyCollection<TSymbol> symbols,
+        ICompletionItemFactory<TSymbol> itemFactory,
         string partialName,
-        CompletionItemKind kind,
-        string detail,
+        bool preferSnippetInsertion,
         List<CompletionItem> target ) where TSymbol : SymbolBase
     {
-        var stringBuilder = new StringBuilder( 256 );
+        if( itemFactory.TryCreateFixedSnippet( partialName, out var fixedSnippetItem ) )
+        {
+            target.Add( fixedSnippetItem );
+        }
 
+        // ReSharper disable once LoopCanBeConvertedToQuery
         foreach( var symbol in symbols )
         {
-            stringBuilder.Clear();
-
-            if( symbol is CallbackSymbol callbackSymbol )
-            {
-                if( TryBuildCallbackSnippetItem( callbackSymbol, partialName, stringBuilder, out var completionItem ) )
-                {
-                    target.Add( completionItem );
-
-                    continue;
-                }
-            }
-
-            if( symbol is CommandSymbol commandSymbol )
-            {
-                var completionItem = BuildCommandItem( commandSymbol );
-
-                target.Add( completionItem );
-
-                continue;
-            }
-
-            if( symbol is VariableSymbol variableSymbol )
-            {
-                var item = BuildVariableItem( variableSymbol, partialName );
-
-                target.Add( item );
-
-                continue;
-            }
-
-            var document = DocumentUtility.GetCommentOrDescriptionText( symbol );
-            document = string.IsNullOrEmpty( document ) ? null : document;
-
-            target.Add(
-                new CompletionItem(
-                    Label: symbol.Name.Value,
-                    Kind: kind,
-                    Detail: detail,
-                    Documentation: document,
-                    InsertTextFormat: InsertTextFormat.PlainText,
-                    InsertText: symbol.Name.Value
-                )
-            );
+            var completionItem = itemFactory.Create( symbol, partialName, preferSnippetInsertion );
+            target.Add( completionItem );
         }
-    }
-
-    private static bool TryBuildCallbackSnippetItem( CallbackSymbol callbackSymbol, string partialName, StringBuilder stringBuilder, out CompletionItem result )
-    {
-        var document = DocumentUtility.GetCommentOrDescriptionText( callbackSymbol );
-        document = string.IsNullOrEmpty( document ) ? null : document;
-
-        result = null!;
-
-        if( !partialName.StartsWith( "on" ) )
-        {
-            return false;
-        }
-
-        if( callbackSymbol.ArgumentCount > 0 )
-        {
-            stringBuilder.Append( "on " ).Append( callbackSymbol.Name.Value ).Append( '(' );
-
-            for( var snippetIndex = 0; snippetIndex < callbackSymbol.ArgumentCount; snippetIndex++ )
-            {
-                stringBuilder.Append( "${" )
-                             .Append( snippetIndex + 1 ).Append( ':' )
-                             .Append( callbackSymbol.Arguments[ snippetIndex ].Name.Value )
-                             .Append( '}' );
-
-                if( snippetIndex < callbackSymbol.ArgumentCount - 1 )
-                {
-                    stringBuilder.Append( ", " );
-                }
-            }
-
-            var codeIndex = callbackSymbol.ArgumentCount + 1;
-
-            stringBuilder.AppendLine( ")" )
-                         .AppendLine( $"    ${{{codeIndex}:code}}" )
-                         .AppendLine( "end on" );
-        }
-        else
-        {
-            stringBuilder.Append( "on " ).AppendLine( callbackSymbol.Name.Value )
-                         .AppendLine( "    ${1:code}" )
-                         .AppendLine( "end on" );
-        }
-
-        result = new CompletionItem(
-            Label: $"on {callbackSymbol.Name.Value}",
-            Kind: CompletionItemKind.Snippet,
-            Detail: "Callback",
-            Documentation: document,
-            InsertTextFormat: InsertTextFormat.Snippet,
-            InsertText: stringBuilder.ToString()
-        );
-
-        return true;
-    }
-
-    private static CompletionItem BuildVariableItem( VariableSymbol variableSymbol, string partialName )
-    {
-        var insertText = variableSymbol.Name.Value;
-        var kind = CompletionItemKind.Variable;
-
-        var document = DocumentUtility.GetCommentOrDescriptionText( variableSymbol );
-        document = string.IsNullOrEmpty( document ) ? null : document;
-
-        if( !variableSymbol.BuiltIn && variableSymbol.Modifier.IsConstant() )
-        {
-            kind = CompletionItemKind.Constant;
-        }
-
-        return new CompletionItem(
-            Label: variableSymbol.Name.Value,
-            Kind: kind,
-            Detail: variableSymbol.BuiltIn ? "Built-in Variable" : "User Variable",
-            Documentation: document,
-            InsertTextFormat: InsertTextFormat.PlainText,
-            InsertText: insertText
-        );
-    }
-
-    private static CompletionItem BuildCommandItem( CommandSymbol commandSymbol )
-    {
-        var document = DocumentUtility.GetCommentOrDescriptionText( commandSymbol );
-        document = string.IsNullOrEmpty( document ) ? null : document;
-
-        return new CompletionItem(
-            Label: commandSymbol.Name.Value,
-            Kind: CompletionItemKind.Method,
-            Detail: "Command",
-            InsertTextFormat: InsertTextFormat.PlainText,
-            Documentation: document,
-            InsertText: commandSymbol.Name.Value
-        );
-    }
-
-    private static CompletionItem BuildCommandSnippet( CommandSymbol commandSymbol, StringBuilder stringBuilder )
-    {
-        var document = DocumentUtility.GetCommentOrDescriptionText( commandSymbol );
-        document = string.IsNullOrEmpty( document ) ? null : document;
-
-        if( commandSymbol.Arguments.Count == 0 )
-        {
-            return new CompletionItem(
-                Label: commandSymbol.Name.Value,
-                Kind: CompletionItemKind.Method,
-                Detail: "Command",
-                Documentation: document,
-                InsertTextFormat: InsertTextFormat.PlainText,
-                InsertText: commandSymbol.Name.Value
-            );
-        }
-
-        stringBuilder.Append( commandSymbol.Name.Value );
-        stringBuilder.Append( "(" );
-
-        var index = 1;
-        var argCount = commandSymbol.Arguments.Count;
-
-        foreach( var arg in commandSymbol.Arguments )
-        {
-            stringBuilder.Append( "${" )
-                         .Append( index ).Append( ":" )
-                         .Append( arg.Name.Value )
-                         .Append( "}" );
-
-            if( index < argCount )
-            {
-                stringBuilder.Append( ", " );
-            }
-
-            index++;
-        }
-
-        stringBuilder.Append( ")" );
-
-        return new CompletionItem(
-            Label: commandSymbol.Name.Value,
-            Kind: CompletionItemKind.Method,
-            Detail: "Command",
-            Documentation: document,
-            InsertTextFormat: InsertTextFormat.Snippet,
-            InsertText: stringBuilder.ToString()
-        );
     }
 }