+++ Compiler/Shared/Shared/Domain/Compilation/Symbols/MetaData/DataTypeUtility.cs
@@ -27,18 +27,6 @@ public static class DataTypeUtility
     public static IReadOnlyCollection<string> KspTypeCharactersAsString { get; }
         = KspTypeCharactersTable.Select( x => x.ToString() ).ToList();
 
-    /// <summary>
-    /// Get KSP data type character mapped to data type flag.
-    /// </summary>
-    public static IReadOnlyDictionary<char, DataTypeFlag> KspTypeCharacterType { get; } = new Dictionary<char, DataTypeFlag>
-    {
-        { '$', DataTypeFlag.TypeInt },
-        { '%', DataTypeFlag.TypeIntArray },
-        { '~', DataTypeFlag.TypeReal },
-        { '?', DataTypeFlag.TypeRealArray },
-        { '@', DataTypeFlag.TypeString },
-        { '!', DataTypeFlag.TypeStringArray }
-    };
     #endregion
 
     /// <summary>
@@ -57,6 +45,42 @@ public static class DataTypeUtility
     public static bool StartsWithDataTypeCharacter( string text )
     {
         return text.Length > 0 && IsDataTypeCharacter( text[ 0 ] );
+    }
+
+    /// <summary>
+    /// Convert to KSP data type character from data type flag.
+    /// </summary>
+    /// <exception cref="ArgumentException">Unknown type or flag has multiple types</exception>
+    public static string GuessKspTypeCharacter( DataTypeFlag dataType )
+    {
+        return !TryGuessKspTypeCharacter( dataType, out var result )
+            ? throw new ArgumentException( $"unknown ksp type (or {nameof( dataType )} has multiple type : {dataType}" )
+            : result;
+    }
+
+    /// <summary>
+    /// Convert to KSP data type character from data type flag.
+    /// </summary>
+    /// <remarks>
+    /// If dataType has multiple type flags, return false.
+    /// </remarks>
+    public static bool TryGuessKspTypeCharacter( DataTypeFlag dataType, out string result )
+    {
+        var unknown = string.Empty;
+
+        // ReSharper disable once SwitchExpressionHandlesSomeKnownEnumValuesWithExceptionInDefault
+        result = dataType switch
+        {
+            DataTypeFlag.TypeInt         => "$",
+            DataTypeFlag.TypeIntArray    => "%",
+            DataTypeFlag.TypeReal        => "~",
+            DataTypeFlag.TypeRealArray   => "?",
+            DataTypeFlag.TypeString      => "@",
+            DataTypeFlag.TypeStringArray => "!",
+            _                            => unknown
+        };
+
+        return result != unknown;
     }
 
     /// <summary>