+++ Tools/SimpleCodeGenerator/simple_codegen.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-pushd `dirname $0` > /dev/null
-this_dir=`pwd`
-popd > /dev/null
-
-uv run python $this_dir/simple_codegen.py "${@}"