+++ Tools/SimpleCodeGenerator/templates/examples/example.cs
@@ -1,9 +1,0 @@
-namespace ${namespace}
-{
-    /// <summary>
-    /// ${description}
-    /// </summary>
-    public class ${classname}
-    {
-    }
-}