+++ Tools/ast-generator/templates/node_id_enum.cs
@@ -1,0 +1,7 @@
+namespace KSPCompiler.Shared.Domain.Compilation.Ast.Nodes;
+
+public enum AstNodeId
+{
+    None,{% for name in names %}
+    {{ name }}{{ "," if not loop.last else "" }}{% endfor %}
+}