+++ Tools/csharp-project-template/cli.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+uv run python main.py templates/cliapp.csproj $1